# backend/internal/api/handlers/openapi.yaml
openapi: 3.0.3
info:
  title: InvestIQ API
  description: |
    Personal financial operating system backend.
    All endpoints except `/health` and `/docs/*` require a Bearer token.
    **Dev mode**: call `GET /auth/dev-login` to get a token; paste it into the Authorize dialog.
    **Clerk mode**: get a session JWT from the Clerk dashboard or frontend.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local dev
  - url: https://moodmarket-production.up.railway.app
    description: Production
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    BrokerageStatus:
      type: object
      properties:
        id:
          type: string
          example: a1b2c3d4e5f6a7b8
        name:
          type: string
          example: Bonds Account
        asset_categories:
          type: array
          items:
            type: string
            enum: [equity, bond, default]
          example: [bond]
        connected:
          type: boolean
        base_url:
          type: string
          example: https://paper-api.alpaca.markets
        connected_at:
          type: string
          format: date-time
    AddBrokerageConnectionRequest:
      type: object
      required: [name, asset_categories, api_key, secret_key]
      properties:
        id:
          type: string
          description: Stable ID — omit to auto-generate
        name:
          type: string
          example: Bonds Account
        asset_categories:
          type: array
          items:
            type: string
            enum: [equity, bond, default]
          example: [bond]
        api_key:
          type: string
          example: PKxxxxxxxxxxxxxxxxxxxxxxxx
        secret_key:
          type: string
          example: xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx
        base_url:
          type: string
          example: https://paper-api.alpaca.markets
...
security:
  - bearerAuth: []
paths:
...
  /brokerage/connect:
    post:
      summary: Connect Alpaca brokerage
      description: Encrypts and stores per-user Alpaca API credentials in MongoDB.
      tags: [Brokerage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [api_key, secret_key]
              properties:
                api_key:
                  type: string
                secret_key:
                  type: string
                base_url:
                  type: string
                  default: https://paper-api.alpaca.markets
      responses:
        '200':
          description: Connection status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrokerageStatus'
    delete:
      summary: Disconnect brokerage
      tags: [Brokerage]
      responses:
        '200':
          description: Disconnected
          content:
            application/json:
              schema:
                type: object
                properties:
                  connected:
                    type: boolean
                    example: false
  /brokerage/connections:
    post:
      summary: Add brokerage connection
      description: Adds or updates a named brokerage connection with asset-category routing rules.
      tags: [Brokerage]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddBrokerageConnectionRequest'
      responses:
        '200':
          description: Connection added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BrokerageStatus'
        '400':
          description: Validation error
  /brokerage/connections/{id}:
    delete:
      summary: Remove brokerage connection
      tags: [Brokerage]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Connection ID from GET /users/profile brokerages array
      responses:
        '204':
          description: Removed
...